+++ src/ast/nodes/CallExpression.ts
@@ -51,7 +51,7 @@ export default class CallExpression extends NodeBase implements DeoptimizableEnt
 					{
 						code: 'EVAL',
 						message: `Use of eval is strongly discouraged, as it poses security risks and may cause issues with minification`,
-						url: 'https://github.com/rollup/rollup/wiki/Troubleshooting#avoiding-eval'
+						url: 'https://rollupjs.org/guide/en#avoiding-eval'
 					},
 					this.start
 				);