+++ src/utils/collapseSourcemaps.ts
@@ -93,7 +93,7 @@ class Link {
 							names.push(traced.name);
 						}
 
-						tracedSegment[4] = nameIndex;
+						(<SourceMapSegmentVector>tracedSegment)[4] = nameIndex;
 					}
 
 					tracedLine.push(tracedSegment);
@@ -177,7 +177,7 @@ export default function collapseSourcemaps(
 					message: `Sourcemap is likely to be incorrect: a plugin${
 						map.plugin ? ` ('${map.plugin}')` : ``
 					} was used to transform files, but didn't generate a sourcemap for the transformation. Consult the plugin documentation for help`,
-					url: `https://github.com/rollup/rollup/wiki/Troubleshooting#sourcemap-is-likely-to-be-incorrect`
+					url: `https://rollupjs.org/guide/en#warning-sourcemap-is-likely-to-be-incorrect`
 				});
 
 				map = {