+++ test/function/samples/default-not-reexported/_config.js
@@ -17,6 +17,6 @@ module.exports = {
 			2:
 			3: console.log( def );
 		`,
-		url: `https://github.com/rollup/rollup/wiki/Troubleshooting#name-is-not-exported-by-module`
+		url: `https://rollupjs.org/guide/en#error-name-is-not-exported-by-module-`
 	}
 };