+++ test/hooks/index.js
@@ -1073,4 +1073,62 @@ module.exports = input;
 				assert.equal(renderErrorCount, 1, 'renderError count');
 			});
 	});
+
+	it('assigns chunk IDs before creating outputBundle chunks', () => {
+		const chunks = [];
+		return rollup
+			.rollup({
+				input: 'input',
+				experimentalCodeSplitting: true,
+				plugins: [
+					loader({
+						input: `export default [import('a'), import('b')];`,
+						a: `import d from 'd'; import c from 'c'; export default () => c();`,
+						b: `import c from 'c'; export default () => c();`,
+						c: `export default () => console.log('c');`,
+						d: `export default {};`
+					}),
+					{
+						renderChunk(code, chunk, options) {
+							chunks.push({
+								fileName: chunk.fileName,
+								imports: chunk.imports,
+								modules: Object.keys(chunk.modules)
+							});
+						}
+					}
+				]
+			})
+			.then(bundle =>
+				bundle.generate({
+					entryFileNames: '[name].js',
+					chunkFileNames: '[name].js',
+					format: 'esm'
+				})
+			)
+			.then(() => {
+				assert.deepEqual(chunks, [
+					{
+						fileName: 'input.js',
+						imports: [],
+						modules: ['input']
+					},
+					{
+						fileName: 'a.js',
+						imports: ['chunk.js'],
+						modules: ['d', 'a']
+					},
+					{
+						fileName: 'chunk.js',
+						imports: [],
+						modules: ['c']
+					},
+					{
+						fileName: 'b.js',
+						imports: ['chunk.js'],
+						modules: ['b']
+					}
+				]);
+			});
+	});
 });